requirements.txt
langchain-chroma>=0.0.1
langchain-google-genai>=0.0.1
langchain-core>=0.1.0
chromadb>=0.4.15
langsmith>=0.1.0
langchain>=0.1.0
python-dotenv>=1.0.0
